fix: fix failing test
