+++ packages/p2p-v2/src/components/PaymentMethodCard/__tests__/PaymentMethodCardHeader.spec.tsx
@@ -43,6 +43,7 @@ describe('PaymentMethodCardHeader', () => {
                 isSelectable
                 onDeletePaymentMethod={() => undefined}
                 onEditPaymentMethod={() => undefined}
+                onSelectPaymentMethod={() => undefined}
                 type='bank'
             />
         );
